requirements.txt
flask>=3.1,<4
authlib>=1.4,<2
python-dotenv>=1,<2
requests>=2.32,<3
gunicorn>=22,<23

sqlalchemy>=2,<3
alembic>=1.13,<2
pymysql>=1.1,<2
